Update padding to for screens larger than 1680px
The mod team alumni div was still positioned incorrectly for me with a browser window larger than 1680px. This changes the rule for .wrapper > .inner, so that screens larger than a max-width of 1680px also have the padding adjusted. As far as I can tell this shouldn't affect smaller screen sizes, since those are taken care of by the other media rules. Similar to #15.